requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
chardet==5.2.0
click==8.1.7
cssselect==1.2.0
Deprecated==1.3.1
diskcache==5.6.3
fastapi==0.115.2
greenlet==3.0.3
h11==0.16.0
html2text==2024.2.26
htmlmin==0.1.12
idna==3.10
Jinja2==3.1.6
limits==5.8.0
lxml==5.3.0
lxml_html_clean==0.4.0
MarkupSafe==3.0.3
packaging==26.0
pillow==11.0.0
playwright==1.47.0
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
python-dotenv==1.0.1
python-multipart==0.0.18
readability-lxml==0.8.1
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.6
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.1.1
SQLAlchemy==2.0.46
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.32.0
wrapt==2.1.1